Adjusted paths to be compatible with eclipse.
All paths to the sounds, pictures, and level files were modified.
To make things easier, the class level fields in the Block class which
granted access to the pictures associated with the different kinds of blocks
were changed from being individual fields to being stored in a class level
HashMap. The Game class, Gui class, and each effected block were adjusted to reflect that change.